// src/shared/types.ts
export type LocationInput = {
  label: string;
  latitude: number;
  longitude: number;
  timeZone?: string;
};